Move user Forms to another file
#104, #105
