+++ accounts/templates/accounts/UpdateUser_profile.html
@@ -8,7 +8,7 @@
     <div class="col-md-8">
         <div class="card">
             <div class="card-header bg-info text-white">
-                User Detail Page
+                 Szczegóły konta
             </div>
             <div class="row p-3">
                 <div class="col-md-12">
@@ -18,8 +18,8 @@
                         {{ p_form|crispy }}
 
                         <div style="text-align: center">
-                            <input class="btn btn-info" type="submit" value="Update">
-                            <a class="btn btn-info"  href="{% url 'accounts:profile'%}">Cancel</a>
+                            <input class="btn btn-info" type="submit" value="Zapisz">
+                            <a class="btn btn-info"  href="{% url 'accounts:profile'%}">Anuluj</a>
                         </div>
                     </form>
                 </div>
